Cleaned up server
